Use libcurl instead of Python
Embedding a python interpreter to talk to pager duty and send emails
doesn't scale well as it's single threaded.

Here we are adding a thread pool based on the multi interace of libcurl so
we can create HTTP and emails clients easily and more performant, as we
can handle thousands of connections per thread.

Get rid of python_interpreter code.